requirements/base_req.txt
pytz==2021.1  # https://github.com/stub42/pytz
python-slugify==5.0.2  # https://github.com/un33k/python-slugify
Pillow==8.2.0  # https://github.com/python-pillow/Pillow
argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi
redis==3.5.3  # https://github.com/andymccurdy/redis-py
celery  # pyup: < 5.0,!=4.4.7  # https://github.com/celery/celery
django-celery-beat # https://github.com/celery/django-celery-beat
flower  # https://github.com/mher/flower
psycopg2-binary==2.9.1 # https://pypi.org/project/psycopg2-binary/

# Django
# ------------------------------------------------------------------------------
django==3.2.4  # pyup: < 3.2  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.1.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.44.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.11.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.12.1  # https://github.com/jazzband/django-redis
django-phonenumber-field[phonenumbers]
djangorestframework-filters
django-filter

# Django API
# ------------------------------------------------------------------------------
django-rest-framework
drf-spectacular
dj-rest-auth==2.1.10

# Testing
# ------------------------------------------------------------------------------
djangorestframework-simplejwt==4.6.0
unittest-xml-reporting==3.0.4
responses
coveralls==1.11.1
pycrunch-engine==1.2.1

# Django debug
# ------------------------------------------------------------------------------
factory-boy==3.2.0  # https://github.com/FactoryBoy/factory_boy
django-debug-toolbar==3.2.1  # https://github.com/jazzband/django-debug-toolbar
django-extensions==3.1.3  # https://github.com/django-extensions/django-extensions
django-coverage-plugin==1.8.0  # https://github.com/nedbat/django_coverage_plugin
pytest-django==4.4.0  # https://github.com/pytest-dev/pytest-django

# Deploy
# ------------------------------------------------------------------------------
gunicorn==20.1.0  # https://github.com/benoitc/gunicorn

# Email
# ------------------------------------------------------------------------------
django-anymail[mailgun]==8.2  # https://github.com/anymail/django-anymail
